split Unterstrasser implementation and tests into separate files
